requirements_Version2.txt
numpy
pandas
scikit-learn
opencv-python
matplotlib
tensorflow
keras
Pillow
tk